Documents/Invensis/requirements.txt
Flask>=2.3.0
Flask-Login>=0.6.0
Flask-Mail>=0.9.0
Werkzeug>=2.3.0
Pillow>=9.0.0
reportlab>=1.5.0
pandas>=1.5.0
python-dotenv>=1.0.0
email-validator>=2.0.0
pymongo>=4.0.0
dnspython>=2.0.0
PyJWT>=2.0.0
openpyxl>=3.1.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
openai>=1.0.0 